templater/requirements.txt
APScheduler==3.6.3
configparser==5.0.0
dnspython==1.16.0
docxtpl==0.10.0
Jinja2==2.10.1
parameterized==0.7.4
PasteDeploy==2.1.0
plaster-pastedeploy==0.7
pymongo==3.10.1
pyramid==1.10.4
pyramid-debugtoolbar==4.6.1
pyramid-jinja2==2.8
python-docx==0.8.10
requests==2.24.0
secretary==0.2.19
selenium==3.141.0
tzlocal==2.1
waitress==1.4.4
xlrd==1.2.0
